removed unnecessary code and added some docs
