+++ include/tools.h
@@ -1,0 +1,6 @@
+#ifndef TOOLS_H
+#define TOOLS_H
+
+void itos(int num, char *buffer);
+
+#endif
